requirements.txt
torch>=1.4.0
torchvision>=0.5.0
pylidc>=0.2.1
matplotlib>=3.1.2
scikit-image>=0.14.2
backpack >= 0.0
ipdb >= 0.0
tqdm >= 0.0
pandas >= 0.0
scikit-learn >= 0.0
backpack-for-pytorch >= 0.0
haven-ai >= 0.0